Move dispatcher logic into router
