+++ cli/interface/generate/controller.js
@@ -34,49 +34,18 @@ module.exports = (function() {
   }
 
   function generateRoute(controllerName, controllerPath, controllerRoute) {
-
-    /* generator: end imports */
-
-    let importStatement = [
-      'const ',
-        controllerName,
-      ' = ',
-      'Nodal.require(\'app/controllers/',
-        controllerPath.join('/'),
-      '\');'
-    ].join('')
-
+    
     let routeStatement = [
       `router.route('/`,
       controllerPath.slice(0, -1).concat([controllerRoute]).join('/'),
-      `/:id', `,
-      controllerName,
-      ');'
+      `/{id}').use('`,
+      controllerPath,
+      `');`
     ].join('');
 
     let routes = fs.readFileSync('./app/router.js').toString();
 
     routes = routes.split('\n');
-
-    let importIndex = routes.map(function(v, i) {
-      return {
-        spaces: v.indexOf('/* generator: end imports */'),
-        index: i
-      }
-    }).filter(function(v) {
-      return v.spaces > -1;
-    }).pop();
-
-    if (importIndex !== undefined) {
-
-      routes = routes.slice(0, importIndex.index - 1).concat(
-        [
-          Array(importIndex.spaces + 1).join(' ') + importStatement,
-        ],
-        routes.slice(importIndex.index - 1)
-      );
-
-    }
 
     let routeIndex = routes.map(function(v, i) {
       return {
